3654
git-svn-id: https://192.168.0.7/svn/team11/eRIMS_SONIC_New/trunk/eRIMS_Sonic@77498 8092c8c6-7ab6-f546-91b1-2980cf571e14

+++ eRims_SonicWeb/DashBoard/DealershipDetail.aspx.cs
@@ -365,7 +365,7 @@ public partial class DealershipDetail : System.Web.UI.Page
 
             if (dsDetail.Tables[0].Rows[0]["Score"] != DBNull.Value)
             {
-                Score = Convert.ToInt32(dsDetail.Tables[0].Rows[0]["Score"]);
+                Score = Convert.ToDecimal(dsDetail.Tables[0].Rows[0]["Score"]);
                 lblSLT_Performance.Text =  dsDetail.Tables[0].Rows[0]["Score"].ToString() + " (" + GetPerformanceLevel_SLT(Score) + ")";
             }
         }
